LayoutTests/svg/css/scientific-numbers-expected.txt

Test scientific numbers on values for SVG presentation attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test default font-size of '16px'
PASS getComputedStyle(text).fontSize is "16px"
PASS getComputedStyle(text).fontSize is "100px"

Test positive exponent values with 'e'
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"

Test positive exponent values with 'E'
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"

Test negative exponent values with 'e'
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"

Test negative exponent values with 'E'
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"

Test negative numbers with exponents
PASS getComputedStyle(text).baselineShift is "-50px"
PASS getComputedStyle(text).baselineShift is "100px"
PASS getComputedStyle(text).baselineShift is "-50px"
PASS getComputedStyle(text).baselineShift is "100px"
PASS getComputedStyle(text).baselineShift is "-50px"
PASS getComputedStyle(text).baselineShift is "100px"

Test if value and 'em' still works
PASS getComputedStyle(text).baselineShift is "50em"
PASS getComputedStyle(text).baselineShift is "100px"

Trailing and leading whitespaces
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "50px"
PASS getComputedStyle(text).fontSize is "100px"

Test behavior on overflow
PASS getComputedStyle(text).baselineShift is "baseline"
PASS getComputedStyle(text).baselineShift is "100px"
PASS getComputedStyle(text).baselineShift is "baseline"
PASS getComputedStyle(text).baselineShift is "100px"

Invalid values
PASS getComputedStyle(text).fontSize is "16px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "16px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "16px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "16px"
PASS getComputedStyle(text).fontSize is "100px"
PASS getComputedStyle(text).fontSize is "16px"
PASS getComputedStyle(text).fontSize is "100px"
PASS successfullyParsed is true

TEST COMPLETE
